requirements.txt
django-bootstrap5
django-phonenumber-field[phonenumberslite]
django-two-factor-auth
fpdf2
pyyaml